requirements_ml.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
scikit-learn
